moved main_test to newly created dbg folder
